fix: refactor warehouse template
