+++ apps/gauzy/src/app/@core/constants/layout.constants.ts
@@ -28,6 +28,7 @@ export enum ComponentEnum {
 	APPROVAL_POLICY = 'APPROVAL_POLICY',
 	PRODUCT_CATEGORY = 'PRODUCT_CATEGORY',
 	PRODUCT_TYPE = 'PRODUCT_TYPE',
+	WAREHOUSE = 'WAREHOUSE',
 	PIPELINE_DEALS = 'PIPELINE_DEALS',
 	TIME_OFF = 'TIME_OFF',
 	TIME_OFF_SETTINGS = 'TIME_OFF_SETTINGS',
