+++ apps/gauzy/src/app/pages/inventory/components/manage-warehouses/warehouse-form/warehouse-form.component.scss
@@ -2,11 +2,6 @@ img {
   width: 100%;
 }
 
-nb-card {
-  width: 800px;
-  max-width: 90%;
-}
-
 .warehouse-image {
   width: fit-content;
   height: fit-content;
