+++ apps/gauzy/src/app/pages/inventory/components/table-inventory-items/table-inventory.component.ts
@@ -172,7 +172,7 @@ export class TableInventoryComponent
 	}
 
 	manageWarehouses() {
-		this.router.navigate(['/pages/organization/inventory/warehouses']);
+		this.router.navigate(['/pages/organization/inventory/warehouses/all']);
 	}
 
 	onAddInventoryItem() {
